+++ pupyl/duplex/file_io.py
@@ -5,15 +5,11 @@ import csv
 import bz2
 import lzma
 import gzip
-import uuid
 import tarfile
 import zipfile
 import tempfile
 import mimetypes
 from io import BytesIO
-from pathlib import Path
-from shutil import rmtree
-from platform import system
 from itertools import cycle
 from enum import Enum, auto
 from datetime import datetime
@@ -23,10 +19,11 @@ from urllib.request import Request, urlopen
 
 import termcolor
 
+from pupyl.addendum.operators import intmul
 from pupyl.duplex.file_types import FileType
+from pupyl.duplex.temporary import SafeTemporaryResource
 from pupyl.duplex.exceptions import FileTypeNotSupportedYet, \
     FileScanNotPossible
-from pupyl.addendum.operators import intmul
 
 
 class Protocols(Enum):
@@ -46,82 +43,6 @@ class Protocols(Enum):
     UNKNOWN = auto()
     HTTP = auto()
     FILE = auto()
-
-
-class SafeTemporaryResource:
-    """Creates temporary files or directories."""
-    def __init__(self, **kwargs):
-        """Defines the temporary resource.
-
-        Parameters
-        ----------
-        file_name: str
-            To define a name for the temporary file. If the file
-            exists, it will try to delete the file before anything else.
-
-        directory_name: str
-            To define a name for the temporary directory. If the directory
-            not exists, it will be created.
-
-        user_defined: bool
-            If should be created a temporary resource based on user's names
-            or not.
-        """
-        self._file_name = kwargs.get('file_name')
-        self._directory_name = kwargs.get('directory_name')
-
-        if system() == 'Windows':
-            default_temp_dir = str(Path.home())
-        else:
-            default_temp_dir = tempfile.gettempdir()
-
-        if kwargs.get('user_defined'):
-            if self._directory_name and not self._file_name:
-                self._temp_path = os.path.join(
-                    default_temp_dir, self._directory_name
-                )
-
-                os.mkdirs(self._temp_path, exists_ok=True)
-            elif not self._directory_name and self._file_name:
-                self._temp_path = os.path.join(
-                    default_temp_dir, self._file_name
-                )
-        else:
-            if system() == 'Windows':
-                self._temp_path = os.path.join(
-                    default_temp_dir, str(uuid.uuid4())
-                )
-
-                os.mkdir(self._temp_path)
-            else:
-                self._temp_dir = tempfile.TemporaryDirectory()
-                self._temp_path = self._temp_dir.name
-
-    @property
-    def name(self):
-        """Getter for property name."""
-        return self._temp_path
-
-    def __str__(self):
-        """Returns the string representation."""
-        return self.name
-
-    def __enter__(self):
-        """Opens the SafeTemporaryResource context."""
-        return self
-
-    def cleanup(self):
-        """Clean up temporary resources."""
-        try:
-            self._temp_dir.cleanup()
-        except AttributeError:
-            rmtree(self._temp_path.name)
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        """Opens the SafeTemporaryResource context."""
-        del exc_type, exc_val, exc_tb
-        
-        self.cleanup()
 
 
 class FileIO(FileType):
@@ -645,8 +566,7 @@ class FileIO(FileType):
         str:
             Paths of the already untarred files on the temporary directory.
         """
-        # Won't explicitly remove the temporary directory
-        temp_dir = tempfile.mkdtemp()
+        temp_dir = SafeTemporaryResource()
 
         inferred_protocol = cls._infer_protocol(uri)
 
@@ -655,7 +575,7 @@ class FileIO(FileType):
                 uri,
                 file_reader.format(stream_type=':')
             ) as tar_file:
-                tar_file.extractall(temp_dir)
+                tar_file.extractall(temp_dir.name)
 
         elif inferred_protocol is Protocols.HTTP:
 
@@ -664,9 +584,9 @@ class FileIO(FileType):
                 mode=file_reader.format(stream_type='|')
             ) as tar_file:
                 for member in cls.progress(tar_file):
-                    tar_file.extract(member, path=temp_dir)
+                    tar_file.extract(member, path=temp_dir.name)
 
-        for root, _, files in os.walk(temp_dir):
+        for root, _, files in os.walk(temp_dir.name):
             for ffile in files:
                 yield os.path.join(root, ffile)
